requirements.txt
Django==5.0
django-hitcount==1.3.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
phonenumbers==8.13.27
Pillow==10.1.0
psycopg2==2.9.9
sqlparse==0.4.4
pipenv==2023.11.15
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
